Add conda nvcc installation to source compile
If users don't have sudo access, or want to have different cuda versions for each conda environment, it can be helpful to install nvcc on a per environment basis. It is non-intuitive which of the many conda package (nvcc, cuda_runtime, etc.) you need to install, so I thought it would be useful to include that info in the source compilation instructions. This is how I installed the package locally, and can upload my conda env file if its wanted :)