bitsandbytes, test existence of triton.ops.matmul_perf_model
This fixes the bitsandbytes to work with newer triton versions
which does not anymore include the triton.ops.matmul_perf_model
which has been moved to triton-langs kernels project.
https://github.com/triton-lang/kernels

Fix is to simply catch the import error if triton is available
but the triton.ops.matmul_perf_model is not. In such case the logic
is implemented in a same way than earlier when the triton
was not available at all.

To my understanding there is currently no a way to install
opt_matmul_perf.py from new tritons kernel project directly with pip,
so in future it could propably be good to include that python class
directly in bitsandbytes.

https://github.com/bitsandbytes-foundation/bitsandbytes/issues/1492

Signed-off-by: Mika Laitio <lamikr@gmail.com>